Refactor code and improve structure for matlab interfaces
The update includes the modification of the relative path in 'robustunwrap_matlab.cpp' and 'raiseerror_matlab.cpp' to a simple reference. This improves the structure and readability of the code.

A new file 'CMakeLists.txt' was added to automate the compilation process for matlab interfaces. This should enhance the usability for developers, making the compiling process easier and faster.

A minor change was also made in the use of 'delete' to 'delete[]' which matches the new[] operator used to allocate 'mag', preventing potential issues with memory deallocation.

Lastly, replaced the specific type mxDouble with auto type to allow for flexibility in handling different data type.